+++ packages/components/src/org.standardnotes.advanced-checklist/config-overrides.js
@@ -1,2 +1,0 @@
-const override = require('../../cra-app.override')
-module.exports = override