+++ packages/components/src/packages/org.standardnotes.advanced-checklist/config-overrides.js
@@ -1,0 +1,2 @@
+const override = require('../cra-app.override')
+module.exports = override